# docs/swagger.yaml
basePath: /
definitions:
  dto.CategoryResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  dto.CreateCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  dto.MeResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  dto.UpdateCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для платформы шахматных курсов
  title: Chess Course API
  version: "1.0"
paths:
  /api/admin/categories:
    post:
      consumes:
      - application/json
      description: Создаёт новую категорию курсов. Требует права администратора.
      parameters:
      - description: Данные для создания категории
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Категория успешно создана
        "400":
          description: Некорректный запрос (проверьте поля)
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен (требуется роль admin)
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Категория с таким названием уже существует
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создать новую категорию
      tags:
      - categories
  /api/admin/categories/{id}:
    delete:
      description: Удаляет категорию по ID. Требует права администратора.
      parameters:
      - description: ID категории
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Категория успешно удалена
        "400":
          description: Неверный ID категории
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен (требуется роль admin)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Категория не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: 'Невозможно удалить: категория содержит курсы'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить категорию
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Обновляет название и/или описание категории. Требует права администратора.
      parameters:
      - description: ID категории
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления категории
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Категория успешно обновлена
        "400":
          description: Некорректный запрос (неверный ID или данные)
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Доступ запрещен (требуется роль admin)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Категория не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить категорию
      tags:
      - categories
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Проверяет email и пароль, возвращает JWT access token.
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT токен
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неверный email или пароль
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вход пользователя
      tags:
      - auth
  /api/auth/me:
    get:
      description: Возвращает данные авторизованного пользователя по JWT токену.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MeResponse'
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Текущий пользователь
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя с email, именем и паролем. Возвращает
        201 при успехе.
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email уже зарегистрирован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация нового пользователя
      tags:
      - auth
  /api/categories:
    get:
      description: Возвращает список всех категорий курсов. Доступно без авторизации.
      produces:
      - application/json
      responses:
        "200":
          description: Список категорий
          schema:
            items:
              $ref: '#/definitions/dto.CategoryResponse'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить все категории
      tags:
      - categories
  /api/categories/{id}:
    get:
      description: Возвращает информацию о категории по её идентификатору.
      parameters:
      - description: ID категории
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Категория найдена
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "400":
          description: Неверный ID категории
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Категория не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить категорию по ID
      tags:
      - categories
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"